Add prop to support condensed view during upload
When uploading hundreds of files at a time, showing all of them
in the list causes a very tall element, and also slows down the
browser due to reactively repainting such a massive list on the
timescales at which updates are occurring.

This commit adds a new prop `maxShow` (integer) to the Upload and
FileUploadList components that limits the max number of files
shown at a time in the list. Additional files are truncated to
an element in the bottom that just shows their count.

This doesn't change the default behavior, which remains to show
the entire list.

Additionally, this fixes a bug where individual file progress
was not being rendered in the progress circle component, and also
condenses the height of each file element in the list to make
it more compact in all cases.