Vuetify 2.0 and Vue 2.6 (#154)
* Failed update attempt

* WIP

* style fixes

* Fix search

* Update pug syntax to avoid a parsing issue

* node-sass fix

* Keep handy vue.config.js changes

* Revert .env

* Fix some unit tests

* wip continued, embrace v-col v-row

* icon fixes and layouts

* Add additional headerwidget to FileManager

Add missing accept prop and rename New Item to Upload

* Fix DataTable and upgrade Vuetify

* Add prop to support condensed view during upload

When uploading hundreds of files at a time, showing all of them
in the list causes a very tall element, and also slows down the
browser due to reactively repainting such a massive list on the
timescales at which updates are occurring.

This commit adds a new prop `maxShow` (integer) to the Upload and
FileUploadList components that limits the max number of files
shown at a time in the list. Additional files are truncated to
an element in the bottom that just shows their count.

This doesn't change the default behavior, which remains to show
the entire list.

Additionally, this fixes a bug where individual file progress
was not being rendered in the progress circle component, and also
condenses the height of each file element in the list to make
it more compact in all cases.

* Add to snippet

* Fix remaining unit tests that were failing for different reasons

* Fix style and pug linting errors

* Fix CircleCI timeout issue

* Remove stylus related dependencies

As Vuetify 2 doesn't use stylus anymore

* Replace usage of v-layout, v-flex with new v-row, v-col

* Bump a minor version

* service not server

* Upgrade vuetify and update unit tests

* Bump major version

* Update readme